Add options to cut off a part of the ODT template
In an ODT file, the only styles that will be saved are the styles which
are applied to some content in the document. If you want to create an
empty ODT document for this plugin, and still want to define styles
which will be applied to the wiki content, you have to write some dummy
text, apply the styles to this text, and surround the text in the
DOKUWIKI-ODT-CUT-START and DOKUWIKI-ODT-CUT-STOP tags. This way, the
styles will be retained in the ODT file, but the dummy text will be
removed on export.